[datamodel] Remove std::compose1 (deprecated) and use data model at compile time
std::*_function classes are deprecated and also cumbersome to use. This
commit replaces it with generic iterators which are much easier to
understand. Furthermore the supported data model version is the one
available at compile time and not a fixed version. Before this commit
only XML files up to version 0.11 were supported.